fix: restore abstract base class and split files
