+++ resilient_logger/sources/resilient_log_source.py
@@ -8,8 +8,7 @@ from django.utils import timezone
 
 from resilient_logger.models import ResilientLogEntry
 from resilient_logger.sources import AbstractLogSource
-from resilient_logger.sources.abstract_log_source import AuditLogDocument
-from resilient_logger.utils import get_resilient_logger_config, value_as_dict
+from resilient_logger.sources.resilient_log_source_entry import ResilientLogSourceEntry
 
 TResilientLogSource = TypeVar("TResilientLogSource", bound="ResilientLogSource")
 
@@ -32,52 +31,6 @@ class StructuredResilientLogEntryData:
 
 
 class ResilientLogSource(AbstractLogSource):
-    class Entry:
-        def __init__(self, log: ResilientLogEntry):
-            self.log = log
-
-        def get_id(self) -> str | int:
-            return self.log.id
-
-        def get_document(self) -> AuditLogDocument:
-            config = get_resilient_logger_config()
-            context = (self.log.context or {}).copy()
-            actor = context.pop("actor", "unknown")
-            operation = context.pop("operation", "MANUAL")
-            target = context.pop("target", "unknown")
-            iso_date = (
-                self.log.created_at.astimezone(datetime.timezone.utc)
-                .isoformat(timespec="milliseconds")
-                .replace("+00:00", "Z")
-            )
-
-            extra = {
-                **context,
-                "source_pk": self.get_id(),
-            }
-
-            return {
-                "@timestamp": iso_date,
-                "audit_event": {
-                    "actor": value_as_dict(actor),
-                    "date_time": iso_date,
-                    "operation": operation,
-                    "origin": config["origin"],
-                    "target": value_as_dict(target),
-                    "environment": config["environment"],
-                    "message": self.log.message,
-                    "level": self.log.level,
-                    "extra": extra,
-                },
-            }
-
-        def is_sent(self) -> bool:
-            return self.log.is_sent
-
-        def mark_sent(self) -> None:
-            self.log.is_sent = True
-            self.log.save(update_fields=["is_sent"])
-
     @classmethod
     def create(
         cls: type[TResilientLogSource], *, level: int, message: Any, context: dict
@@ -88,7 +41,7 @@ class ResilientLogSource(AbstractLogSource):
             context=context,
         )
 
-        return ResilientLogSource.Entry(entry)
+        return ResilientLogSourceEntry(entry)
 
     @classmethod
     def bulk_create(
@@ -99,7 +52,7 @@ class ResilientLogSource(AbstractLogSource):
             for obj in objs
         )
 
-        return [ResilientLogSource.Entry(entry) for entry in entries]
+        return [ResilientLogSourceEntry(entry) for entry in entries]
 
     @classmethod
     def create_structured(
@@ -147,7 +100,7 @@ class ResilientLogSource(AbstractLogSource):
     @transaction.atomic
     def get_unsent_entries(
         self, chunk_size: int
-    ) -> Iterator["ResilientLogSource.Entry"]:
+    ) -> Iterator["ResilientLogSourceEntry"]:
         entries = (
             ResilientLogEntry.objects.filter(is_sent=False)
             .order_by("created_at")
@@ -155,7 +108,7 @@ class ResilientLogSource(AbstractLogSource):
         )
 
         for entry in entries:
-            yield ResilientLogSource.Entry(entry)
+            yield ResilientLogSourceEntry(entry)
 
     @transaction.atomic
     def clear_sent_entries(self, days_to_keep: int = 30) -> list[str]:
