+++ resilient_logger/targets/abstract_log_target.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from resilient_logger.sources import AbstractLogSource
+from resilient_logger.sources.abstract_log_source_entry import AbstractLogSourceEntry
 
 
 class AbstractLogTarget(ABC):
@@ -9,5 +9,5 @@ class AbstractLogTarget(ABC):
         raise NotImplementedError()
 
     @abstractmethod
-    def submit(self, entry: AbstractLogSource.Entry) -> bool:
+    def submit(self, entry: AbstractLogSourceEntry) -> bool:
         raise NotImplementedError()
